portfolio-backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
mysqlclient>=2.2
django-cors-headers>=4.3
Pillow>=10.0
